Moved BoxedLink into its own file
